build: refactor `build.gradle.kts` and update dependencies
This commit significantly refactors the build.gradle.kts file to incorporate usage of multiple plugins like Gradle IntelliJ Plugin, Gradle Changelog Plugin, Gradle Qodana Plugin, and Gradle Kover Plugin, which were previously not in use. Using these plugins adds utility and optimizes the build process.

It also updates project properties and versions to be fetched dynamically which adds flexibility to the build and release process. It models changes in the environment settings that are more preferably used than the System.getenv invocations.

Further, the Gradle version is updated to 8.2.1 from 8.1.1 improving the build's performance and feature availability.

A section to configure UI tests plugin is added and lastly, settings.gradle.kts is simplified to include just the rootProject name.

In essence, this commit restructures the build file to make it more robust, efficient and maintainable.