ci: add Dependabot configuration file for Github Actions and Gradle
Created a new configuration file '.github/dependabot.yml' to use Dependabot for managing dependencies in GitHub Actions and Gradle. This configuration checks for updates daily, assigns and notifies "nekofar", labels the pull requests as "dependencies", and sets an open pull request limit of 20. All the pull requests are targeted to the 'develop' branch.