ci: add funding information to the project
A new FUNDING.yml file was created in the .github directory. This includes the necessary details to display the funding button on the project page. The purpose of these changes is to provide a clear path for potential sponsors. It specifies both a Github and a custom funding option for potential sponsors to contribute to the project.