ci: add stale bot configuration file
A new `.github/stale.yml` configuration file for stale bot has been added. The configuration declares that an issue becomes stale after 60 days of inactivity, and it will be closed after a further 7 days of inactivity. Exceptions have been made for issues labeled as `pinned` or `security` where they are never considered stale. Moreover, an issue is marked as stale with a `wontfix` label and a comment is posted to notify about the status change. No comment is posted when closing a stale issue.